+++ Camera-TFT/scripts/setup/servoctrl_setup.sh
@@ -13,8 +13,21 @@ sudo apt-get install -y git-core
 git clone git://git.drogon.net/wiringPi /tmp/rpi-wiringpi
 cd /tmp/rpi-wiringpi
 git pull origin
+
+# Build/install and check if finished successfully
 ./build
-cd $shdir
+if [[ $? -ne 0 ]]; then
+    exit 1
+fi
+
+# Clean-up
+cd ~/
 rm -rf /tmp/rpi-wiringpi
 
+#   [...not finished...]
 
+# Set up System V init script to provide service
+sudo cp $shdir/../service/servoctrl_service.sh /etc/init.d/servoctrl
+sudo chmod 755 /etc/init.d/servoctrl
+
+exit 0
